+++ js/thumbnails.js
@@ -1,0 +1,12 @@
+const getPictureTemplate = ({ url, comments, likes }) => `<a href="#" class="picture">
+<img class="picture__img" src="${url}" width="182" height="182" alt="Случайная фотография">
+<p class="picture__info">
+  <span class="picture__comments">${comments.length}</span>
+  <span class="picture__likes">${likes}</span>
+</p>
+</a>`;
+
+const mainContainer = document.querySelector('.js-pictures');
+const createThumbnails = (data) => mainContainer.insertAdjacentHTML('beforeend', data.map((photo) => getPictureTemplate(photo)).join(''));
+
+export { createThumbnails };
